more language server reorganization
