+++ crates/language-server/src/functionality/capabilities.rs
