+++ crates/language-server/src/functionality/hover.rs
@@ -1,11 +1,9 @@
 use std::sync::Arc;
 
 use common::{input::IngotKind, InputDb};
-use hir::{LowerHirDb, SpannedHirDb};
-use hir_analysis::{
-    name_resolution::{EarlyResolvedPath, NameRes},
-    HirAnalysisDb,
-};
+use hir::LowerHirDb;
+use hir_analysis::{name_resolution::EarlyResolvedPath, HirAnalysisDb};
+use lsp_types::Hover;
 use tracing::info;
 
 use salsa::Snapshot;
@@ -15,9 +13,10 @@ use tower_lsp::jsonrpc::Result;
 use crate::{
     backend::db::LanguageServerDatabase,
     backend::workspace::{IngotFileContext, Workspace},
-    goto::{goto_enclosing_path, Cursor},
-    util::{to_lsp_location_from_scope, to_offset_from_position},
+    util::to_offset_from_position,
 };
+
+use super::goto::{goto_enclosing_path, Cursor};
 
 pub async fn hover_helper(
     db: Snapshot<LanguageServerDatabase>,
@@ -100,61 +99,4 @@ pub async fn hover_helper(
         range: None,
     };
     Ok(Some(result))
-}
-
-use lsp_types::{GotoDefinitionParams, GotoDefinitionResponse, Hover};
-
-pub async fn goto_helper(
-    db: Snapshot<LanguageServerDatabase>,
-    workspace: Arc<RwLock<Workspace>>,
-    params: GotoDefinitionParams,
-) -> Result<Option<GotoDefinitionResponse>> {
-    let workspace = workspace.read().await;
-    // Convert the position to an offset in the file
-    let params = params.text_document_position_params;
-    let file_text = std::fs::read_to_string(params.text_document.uri.path()).ok();
-    let cursor: Cursor = to_offset_from_position(params.position, file_text.unwrap().as_str());
-
-    // Get the module and the goto info
-    let file_path = params.text_document.uri.path();
-    let top_mod = workspace
-        .top_mod_from_file_path(db.as_lower_hir_db(), file_path)
-        .unwrap();
-    let goto_info = goto_enclosing_path(&db, top_mod, cursor);
-
-    // Convert the goto info to a Location
-    let scopes = match goto_info {
-        Some(EarlyResolvedPath::Full(bucket)) => {
-            bucket.iter().map(NameRes::scope).collect::<Vec<_>>()
-        }
-        Some(EarlyResolvedPath::Partial {
-            res,
-            unresolved_from: _,
-        }) => {
-            vec![res.scope()]
-        }
-        None => return Ok(None),
-    };
-
-    let locations = scopes
-        .iter()
-        .filter_map(|scope| *scope)
-        .map(|scope| to_lsp_location_from_scope(scope, db.as_spanned_hir_db()))
-        .collect::<Vec<_>>();
-
-    let _errors = scopes
-        .iter()
-        .filter_map(|scope| *scope)
-        .map(|scope| to_lsp_location_from_scope(scope, db.as_spanned_hir_db()))
-        .filter_map(std::result::Result::err)
-        .map(|err| err.to_string())
-        .collect::<Vec<_>>()
-        .join("\n");
-
-    Ok(Some(lsp_types::GotoDefinitionResponse::Array(
-        locations
-            .into_iter()
-            .filter_map(std::result::Result::ok)
-            .collect(),
-    )))
 }
