+++ crates/language-server/src/functionality/mod.rs
@@ -1,0 +1,6 @@
+mod capabilities;
+pub(super) mod diagnostics;
+pub(super) mod goto;
+pub(super) mod handlers;
+pub(super) mod hover;
+pub(crate) mod streams;
