+++ crates/language-server/src/functionality/streams.rs
